ALTAPPS-206: move profile settings feature to separate package
