+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/injection/AndroidAppComponent.kt
@@ -13,9 +13,9 @@ import org.hyperskill.app.home.injection.HomeComponent
 import org.hyperskill.app.home.injection.PlatformHomeComponent
 import org.hyperskill.app.main.injection.PlatformMainComponent
 import org.hyperskill.app.profile.injection.PlatformProfileComponent
-import org.hyperskill.app.profile.injection.PlatformProfileSettingsComponent
+import org.hyperskill.app.profile_settings.injection.PlatformProfileSettingsComponent
 import org.hyperskill.app.profile.injection.ProfileComponent
-import org.hyperskill.app.profile.injection.ProfileSettingsComponent
+import org.hyperskill.app.profile_settings.injection.ProfileSettingsComponent
 import org.hyperskill.app.step.injection.PlatformStepComponent
 import org.hyperskill.app.step.injection.StepComponent
 import org.hyperskill.app.step_quiz.injection.PlatformStepQuizComponent
