--- shared/src/commonMain/kotlin/org/hyperskill/app/profile/domain/model/Theme.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.profile.domain.model
-
-enum class Theme {
-    LIGHT,
-    DARK,
-    SYSTEM
-}
