+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile_settings/cache/ProfileSettingsCacheKeyValues.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.profile.cache
+package org.hyperskill.app.profile_settings.cache
 
 object ProfileSettingsCacheKeyValues {
     const val PROFILE_SETTINGS = "profile_settings"
