+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile_settings/data/source/ProfileSettingsCacheDataSource.kt
@@ -1,7 +1,7 @@
-package org.hyperskill.app.profile.data.source
+package org.hyperskill.app.profile_settings.data.source
 
-import org.hyperskill.app.profile.domain.model.ProfileSettings
-import org.hyperskill.app.profile.domain.model.Theme
+import org.hyperskill.app.profile_settings.domain.model.ProfileSettings
+import org.hyperskill.app.profile_settings.domain.model.Theme
 
 interface ProfileSettingsCacheDataSource {
     fun getProfileSettings(): Result<ProfileSettings>
