+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile_settings/domain/model/Theme.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.profile_settings.domain.model
+
+enum class Theme {
+    LIGHT,
+    DARK,
+    SYSTEM
+}
