+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile_settings/injection/ProfileSettingsComponentImpl.kt
@@ -1,0 +1,26 @@
+package org.hyperskill.app.profile_settings.injection
+
+import org.hyperskill.app.core.injection.AppGraph
+import org.hyperskill.app.profile_settings.presentation.ProfileSettingsFeature
+import org.hyperskill.app.profile_settings.cache.ProfileSettingsCacheDataSourceImpl
+import org.hyperskill.app.profile_settings.data.repository.ProfileSettingsRepositoryImpl
+import org.hyperskill.app.profile_settings.data.source.ProfileSettingsCacheDataSource
+import org.hyperskill.app.profile_settings.domain.interactor.ProfileSettingsInteractor
+import org.hyperskill.app.profile_settings.domain.repository.ProfileSettingsRepository
+import ru.nobird.app.presentation.redux.feature.Feature
+
+class ProfileSettingsComponentImpl(private val appGraph: AppGraph) : ProfileSettingsComponent {
+    private val profileSettingsCacheDataSource: ProfileSettingsCacheDataSource =
+        ProfileSettingsCacheDataSourceImpl(
+            appGraph.commonComponent.json,
+            appGraph.commonComponent.settings
+        )
+
+    private val profileSettingsRepository: ProfileSettingsRepository =
+        ProfileSettingsRepositoryImpl(profileSettingsCacheDataSource)
+    private val profileSettingsInteractor: ProfileSettingsInteractor =
+        ProfileSettingsInteractor(profileSettingsRepository)
+
+    override val profileSettingsFeature: Feature<ProfileSettingsFeature.State, ProfileSettingsFeature.Message, ProfileSettingsFeature.Action>
+        get() = ProfileSettingsFeatureBuilder.build(profileSettingsInteractor)
+}
