+++ shared/src/iosMain/kotlin/org/hyperskill/app/core/injection/AppGraphImpl.kt
@@ -17,8 +17,8 @@ import org.hyperskill.app.profile.injection.ProfileDataComponent
 import org.hyperskill.app.profile.injection.ProfileDataComponentImpl
 import org.hyperskill.app.profile.injection.ProfileComponent
 import org.hyperskill.app.profile.injection.ProfileComponentImpl
-import org.hyperskill.app.profile.injection.ProfileSettingsComponent
-import org.hyperskill.app.profile.injection.ProfileSettingsComponentImpl
+import org.hyperskill.app.profile_settings.injection.ProfileSettingsComponent
+import org.hyperskill.app.profile_settings.injection.ProfileSettingsComponentImpl
 import org.hyperskill.app.step.injection.StepComponent
 import org.hyperskill.app.step.injection.StepComponentImpl
 import org.hyperskill.app.step_quiz.injection.StepQuizComponent
