--- shared/src/iosMain/kotlin/org/hyperskill/app/profile/injection/PlatformProfileSettingsComponent.kt
@@ -1,3 +1,0 @@
-package org.hyperskill.app.profile.injection
-
-actual interface PlatformProfileSettingsComponent
